requirements.txt
scapy==2.4.5
numpy==1.21.6
pandas==1.3.5
scikit-learn==1.0.2
xgboost==1.5.2
tensorflow==2.11.0
flask==2.0.3
pyyaml==6.0
joblib==1.1.0
matplotlib==3.5.3
seaborn==0.11.2
threading==0.0.0
# Core dependencies
python>=3.8
scapy==2.4.5
numpy==1.21.6
pandas==1.3.5
scikit-learn==1.0.2
tensorflow==2.11.0
flask==2.0.3

# Additional ML libraries
xgboost==1.5.2
joblib==1.1.0

# Visualization and monitoring
matplotlib==3.5.3
seaborn==0.11.2

# Utilities
pyyaml==6.0
requests==2.28.2

# Testing
pytest==7.2.2
pytest-cov==4.0.0
